+++ NEWS.md
@@ -4,6 +4,7 @@
 * Use `set.seed()` instead of `withr::local_seed()` (#67, @shikokuchuo).
 * Add a vignette to showcase integration with Shiny, co-authored with @dwoodie and with helpful suggestions by @shikokuchuo.
 * Add a new `pad_daemons` argument to routers and controllers to avoid too frequent calls to `mirai::daemons()`.
+* Drop `rlang::duplicate()` (#72, @shikokuchuo).
 
 # crew 0.1.1